Refactored new job post  * moved new job post code into new job details  * deleted new job post  * restructured new job details and progress folder structure  * replaced purple color from theme
