+++ frontend/src/components/new_job_post/new_job_details/Form.jsx
@@ -1,4 +1,4 @@
-import Title from "./Title";
+import Title from "../Title";
 
 const Form = () => {
   return (
