+++ frontend/src/components/new_job_post/progress_bar/CircleCheckComplete.jsx
